Phase #5 Development. Add plugin UI, refactor UX, features, and code structure

+++ rust/src/audio/event.rs
@@ -1,6 +1,6 @@
 #[derive(Clone, Copy, Debug)]
 pub struct PlaybackPosition {
-    pub samples: u64,
+    pub samples: u32,
     pub beat: usize,
     pub bar: usize,
     pub tempo: f32, // Useful for Flutter to interpolate movement
